refactor: 신발 생성 응답객체의 이름을 SneakerIdResponse 로 변경하라
