+++ src/main/java/prgrms/neoike/service/SneakerService.java
@@ -8,15 +8,17 @@ import prgrms.neoike.domain.sneaker.SneakerImage;
 import prgrms.neoike.domain.sneaker.SneakerStock;
 import prgrms.neoike.repository.SneakerRepository;
 import prgrms.neoike.repository.SneakerStockRepository;
+import prgrms.neoike.service.dto.sneaker.SneakerDetailDto;
+import prgrms.neoike.service.dto.sneaker.SneakerDetailResponse;
+import prgrms.neoike.service.dto.sneaker.SneakerIdResponse;
 import prgrms.neoike.service.dto.sneaker.SneakerRegisterDto;
-import prgrms.neoike.service.dto.sneaker.SneakerResponse;
-import prgrms.neoike.service.mapper.SneakerConverter;
+import prgrms.neoike.service.converter.SneakerConverter;
 
 import javax.persistence.EntityExistsException;
 import java.util.List;
 
 import static java.text.MessageFormat.format;
-import static prgrms.neoike.service.mapper.SneakerConverter.*;
+import static prgrms.neoike.service.converter.SneakerConverter.*;
 
 @Service
 @RequiredArgsConstructor
@@ -26,13 +28,13 @@ public class SneakerService {
     private final SneakerStockRepository sneakerStockRepository;
 
     @Transactional
-    public SneakerResponse registerSneaker(SneakerRegisterDto registerDto) {
+    public SneakerIdResponse registerSneaker(SneakerRegisterDto registerDto) {
         validateDuplicatedSneaker(registerDto.sneakerDto().code());
 
         Sneaker retrievedSneaker = saveSneaker(registerDto);
         saveSneakerStocks(registerDto, retrievedSneaker);
 
-        return toSneakerResponse(retrievedSneaker.getId(), retrievedSneaker.getCode());
+        return toSneakerIdResponse(retrievedSneaker.getId(), retrievedSneaker.getCode());
     }
 
     private void saveSneakerStocks(SneakerRegisterDto registerDto, Sneaker retrievedSneaker) {
