+++ src/main/java/prgrms/neoike/service/dto/sneaker/SneakerIdResponse.java
@@ -1,6 +1,6 @@
 package prgrms.neoike.service.dto.sneaker;
 
-public record SneakerResponse(
+public record SneakerIdResponse(
     Long sneakerId,
     String code
 ) {
